refactor!: move clipboard and elements files
Move the "clipboard.lua" and "elements.lua" files from the actions
folder into the parent drex folder. Their content is way to broad and
used throughout all parts of the plugin. Therefore it feels "wrong" to
classify them as actions
