+++ lua/drex/fs.lua
@@ -158,7 +158,7 @@ function M.watch_directory(buffer, path)
                     end
                 end
 
-                local clipboard = require('drex.actions.clipboard').clipboard
+                local clipboard = require('drex.clipboard').clipboard
                 for element, _ in pairs(clipboard) do
                     if vim.startswith(element, path) then
                         clipboard[element] = nil
@@ -185,7 +185,7 @@ function M.watch_directory(buffer, path)
             end
 
             -- check clipboard for elements that have been renamed or deleted outside of Neovim
-            local clipboard = require('drex.actions.clipboard').clipboard
+            local clipboard = require('drex.clipboard').clipboard
             for element, _ in pairs(clipboard) do
                 if vim.startswith(element, path) then
                     if not luv.fs_lstat(element) then
